requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.2.0
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
discord.py==1.4.1
flake8==3.8.3
idna==2.10
mccabe==0.6.1
mojang==0.1.1
multidict==4.7.6
pycodestyle==2.6.0
pyflakes==2.2.0
requests==2.24.0
toml==0.10.1
urllib3==1.25.10
yarl==1.5.1
